Move get_args to its own module
