+++ src/lib.rs
@@ -6,3 +6,8 @@
  * Parses the argument list and execute the given sub-command
  */
 pub mod cli;
+
+/**
+ * Module to retrieve arguments from stdin or config file.
+ */
+pub mod get_args;
